requirements.txt
beautifulsoup4==4.13.3
docx==0.2.4
edapi==0.1.0
faiss-cpu
fastapi==0.115.8
fitz==0.0.1.dev2
langchain==0.3.18
langchain_community==0.3.17
langchain_core==0.3.34
langchain_huggingface==0.1.2
numpy==2.2.2
passlib==1.7.4
pydantic==2.10.6
python-dotenv==1.0.1
python-docx
sentence_transformers==3.4.1
starlette==0.45.3
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
